Inline ImportPath strings
