+++ codegen/config_test.go
@@ -1,10 +1,12 @@
 package codegen
 
 import (
+	"go/build"
 	"os"
 	"path/filepath"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -50,5 +52,29 @@ func TestLoadDefaultConfig(t *testing.T) {
 		cfg, err = LoadDefaultConfig()
 		require.NoError(t, err)
 		require.Equal(t, cfg.SchemaFilename, "schema.graphql")
+	})
+}
+
+func Test_fullPackageName(t *testing.T) {
+	origBuildContext := build.Default
+	defer func() { build.Default = origBuildContext }()
+
+	t.Run("gopath longer than package name", func(t *testing.T) {
+		p := PackageConfig{Filename: "/b/src/y/foo/bar/baz.go"}
+		build.Default.GOPATH = "/a/src/xxxxxxxxxxxxxxxxxxxxxxxx:/b/src/y"
+		var got string
+		ok := assert.NotPanics(t, func() { got = p.ImportPath() })
+		if ok {
+			assert.Equal(t, "/b/src/y/foo/bar", got)
+		}
+	})
+	t.Run("stop searching on first hit", func(t *testing.T) {
+		p := PackageConfig{Filename: "/a/src/x/foo/bar/baz.go"}
+		build.Default.GOPATH = "/a/src/x:/b/src/y"
+		var got string
+		ok := assert.NotPanics(t, func() { got = p.ImportPath() })
+		if ok {
+			assert.Equal(t, "/a/src/x/foo/bar", got)
+		}
 	})
 }
